requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
openai>=1.6.0
spacy>=3.7.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
langdetect>=1.0.9
python-dotenv>=1.0.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.25.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
